ref(layouts): Move reusable layout components
- Moved reusable layout components (e.g., `ArticleBackground`,
  `InstagramEmbed`, and `InstagramPosts`) to "components/layout"
