+++ src/components/layout/Hero.tsx
