+++ src/layout/ErrorPage.tsx
@@ -1,5 +1,5 @@
-import Header from "./Header";
-import Footer from "./Footer";
+import Header from "@/components/layout/Header";
+import Footer from "@/components/layout/Footer";
 
 interface ErrorPageProps {
   errno?: number;
