+++ src/pages/ArticlePage.tsx
@@ -4,11 +4,11 @@ import ReactDOMServer from 'react-dom/server';
 import { ExternalLinkIcon } from 'lucide-react';
 import { Badge } from '@/components/ui/badge';
 
-import RootLayout from '@/layout/Root';
-import Header from '@/layout/Header';
-import Footer from '@/layout/Footer';
+import RootLayout from '@/components/layout/Root';
+import Header from '@/components/layout/Header';
+import Footer from '@/components/layout/Footer';
 import Layout from '@/lib/Layout';
-import { MobileTableOfContents, TableOfContents } from '@/layout/TOC';
+import { MobileTableOfContents, TableOfContents } from '@/components/layout/TOC';
 import { getMarkdownContent, ParsedMarkdown } from '@/lib/markparser';
 import CreateErrorPage from '@/layout/ErrorPage';
 import Article from '@/lib/Article';
