+++ src/pages/CodeOfConduct.tsx
@@ -1,8 +1,8 @@
 import { useEffect, useMemo, useState } from "react";
 import { Navigate, useLocation, useParams } from "react-router-dom";
 import { ExternalLinkIcon } from "lucide-react";
-import Header from "@/layout/Header";
-import Footer from "@/layout/Footer";
+import Header from "@/components/layout/Header";
+import Footer from "@/components/layout/Footer";
 import {
   code_of_conduct as codeOfConduct,
   security_policy as securityPolicy,
