+++ src/pages/Home.tsx
@@ -1,7 +1,7 @@
-import RootLayout from "@/layout/Root";
+import RootLayout from "@/components/layout/Root";
 import HomeLayout from "@/layout/Home";
-import Header from "@/layout/Header";
-import Footer from "@/layout/Footer";
+import Header from "@/components/layout/Header";
+import Footer from "@/components/layout/Footer";
 
 export default function Home() {
   return (
